% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run2048.R
\name{run2048}
\alias{run2048}
\title{The 2048 Game}
\usage{
run2048()
}
\description{
The adjacent numbers can be added if they are equal. When the maximum number reaches 2048, you win.
}
\examples{
run2048()

}
