% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfood.R
\name{runfood}
\alias{runfood}
\title{What to Eat Today}
\usage{
runfood()
}
\description{
A small program is used to decide what to eat today, and currently only contains delicious Guangdong dishes.
}
\examples{
runfood()

}
